Change type of global variable tsteps_per_hour

+++ code/global.cpp
@@ -122,7 +122,7 @@ bool Global::work_stealing        = false;
 bool Global::stop_on_cc_err       = false;
 struct tm* Global::ts_start_tm    = NULL;
 struct tm* Global::ts_end_tm      = NULL;
-int Global::tsteps_per_hour       = 1;
+unsigned int Global::tsteps_per_hour = 1;
 unsigned long Global::expansion_scenario_id = 0;
 float Global::time_step_size_in_h   = 0.0;
 bool  Global::break_sac_loop_if_limit_reached = true;
@@ -635,7 +635,7 @@ void Global::set_tsteps_per_hour(int tsteps_per_hour) {
             cerr << "tsteps_per_hour should be set to a value < 0. This is not allowed." << endl;
             return;
         }
-        Global::tsteps_per_hour = tsteps_per_hour;
+        Global::tsteps_per_hour = (unsigned int) tsteps_per_hour; // safe with the above check
         Global::tsteps_per_hour_init = true;
         // set set_tsteps_per_hour accordingly
         Global::time_step_size_in_h = 1.0f / (float)tsteps_per_hour;
